requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
supabase==1.2.0
django-storages==1.14.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0